chore(tests): Using faker to generate test data
Subtitle: How to get rid of Jean-Pierre DUPONT and toto.

We have in many of our test files random string and sentences. This
shouldn't be up the the developers to figure out those random values.

I've added faker in order to generate those values. Even though faker is
usually used to genretate "trully random" values a seed can be set in
order to have repeatable results.