feat(ARCH-661): add typings to react-components (#4423)
Add the bare minimum configuration and tools to generate typings in @talend/react-components package.

It means from this release react-components will contains a lib/index.d.ts in the release.

You can still use declare module '@talend/react-components'; in your package but you should try to remove it to see if current typings are working well. Here it has been possible for datagrid but not for dataviz.

This came from the fact the new types are inferred by typescript from the javascript code so this is far from perfect.
But with this we will be now able to use ts and tsx files in place of javascript files and so be able to create correct typings.